restructured folder feature, deleted unnecessary files and folders after the reorganization, fixed imports and broken parts and paths
