--- frontend/src/features/container-details/index.ts
@@ -1,2 +1,0 @@
-// Export feature components
-export { default as ContainerDetailsPage } from './ContainerDetailsPage';
