+++ frontend/src/features/container-management/index.ts
@@ -1,1 +1,5 @@
-export { default as ContainerManagementDashboard } from './ContainerManagementDashboard';
+export { default as ContainerManagementDashboard } from './pages/ContainerManagementDashboard';
+// export { default as ContainerManagementPage } from './pages/ContainerManagementPage';
+export { default as ContainerListPage } from './pages/ContainerListPage';
+export { default as ContainerDetailsPage } from './pages/ContainerDetailsPage';
+
