+++ frontend/src/features/container-management/pages/ContainerDetailsPage.tsx
@@ -20,15 +20,15 @@ import containerService, {
   ContainerDetail,
   ContainerMetrics,
   TimeRangeOption,
-} from '../../services/containerService';
-import { ContainerOverviewInfo } from '../../shared/components/ui/Container';
-import { StatusChipActive } from '../../shared/components/ui/StatusChip';
+} from '../../../services/containerService';
+import { ContainerOverviewInfo } from '../../../shared/components/ui/Container';
+import { StatusChipActive } from '../../../shared/components/ui/StatusChip';
 // Import components
-import { TabGroup } from '../../shared/components/ui/Tab';
-import ContainerInfoSection from './ContainerInfoSection';
-import CropsSection from './CropsSection';
+import { TabGroup } from '../../../shared/components/ui/Tab';
+import ContainerInfoSection from '../sections/ContainerInfoSection';
+import CropsSection from '../../container-management/sections/CropsSection';
 // Import feature components
-import MetricsPanel from './MetricsPanel';
+import MetricsPanel from '../widgets/MetricsPanel';
 
 // Define tab options
 type TabOption = 'overview' | 'environment' | 'inventory' | 'devices';
