--- frontend/src/features/containers/index.ts
@@ -1,2 +1,0 @@
-export { default as ContainerListPage } from './ContainerListPage';
-export { default as CreateContainerModal } from './CreateContainerModal';
