+++ frontend/src/router/AppRouter.tsx
@@ -3,10 +3,10 @@ import { Navigate, Route, Routes } from 'react-router-dom';
 
 import { Box } from '@mui/material';
 
-import { ContainerDetailsPage } from '../features/container-details';
+import ContainerDetailsPage from '../features/container-management/pages/ContainerDetailsPage';
 import { ContainerManagementDashboard } from '../features/container-management';
 // Pages
-import { ContainerListPage } from '../features/containers';
+import  ContainerListPage  from '../features/container-management/pages/ContainerListPage';
 import { ExampleFeature } from '../features/example';
 import { ContainerMetricsPage } from '../features/metrics';
 // Components
